Revamp categories and validity checks
Previously it was possible to rate ANY object. This is not a good plan,
as malicious users could rate objects that should not be rated, such as
adminstrative users. This could potentially allow users to see private
data.

Now, models that are to be rated need to be explicitly mentioned as such
in the `settings.py`. This is explained in the updated documentation.

Additionally, category choices were stored as integer choice fields. The
integers for each choice were assigned sequentially as choices were
processed. This would have led to unstable behaviour regarding choices,
as choices could be added and removed in the future, disrupting the
sequence of choices.

This has been fixed by changing the `category` field to a CharField,
instead of the previous integer field. Only predefined values from
`settings.py` are ever inserted into this field, so injection attacks
are not an issue here.